+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Line wrap the file at 100 chars.                                              Th
 
 - Fix the potential reconnect loop in GUI, triggered by the timeout when receiving
   the initial state of the daemon.
+- Fix the bug which caused the account token history to remain stale after logout.
 
 #### Linux
 - Fix startup failure when network device with a hardware address that's not a MAC address is